src/main/java/br/com/infnet/InfoNet/Service/FuncionarioService.java
package br.com.infnet.InfoNet.Service;

import br.com.infnet.InfoNet.Model.Funcionario;
import br.com.infnet.InfoNet.Repository.FuncionarioRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class FuncionarioService {

    @Autowired
    private FuncionarioRepository funcionarioRepository;

    public List<Funcionario> getAllFuncionarios() {
        return funcionarioRepository.findAll();
    }

    public Funcionario getFuncionarioById(Long id) {
        Optional<Funcionario> funcionario = funcionarioRepository.findById(id);
        return funcionario.orElse(null);
    }

    public Funcionario createFuncionario(Funcionario funcionario) {
        return funcionarioRepository.save(funcionario);
    }

    public Funcionario updateFuncionario(Long id, Funcionario funcionario) {
        Optional<Funcionario> funcionarioExistente = funcionarioRepository.findById(id);
        if (funcionarioExistente.isPresent()) {
            Funcionario atualizacao = funcionarioExistente.get();
            atualizacao.setNome(funcionario.getNome());
            atualizacao.setCargo(funcionario.getCargo());
            atualizacao.setSalario(funcionario.getSalario());
            return funcionarioRepository.save(atualizacao);
        }
        return null;
    }

    public void deleteFuncionario(Long id) {
        funcionarioRepository.deleteById(id);
    }
}
